Create preprocess.py
The file contains pre-processing functions such as TimeStamp regularization, resampling, and clipping which are very relevant and useful in today's world. These are some commonly used functions when it comes to processing industrial data.  Once included it would be very helpful to people all over the world especially Industrial analysts who work with time-series data on a regular basis(Examples included forecasting the health of sensors or components in the future with current data)